feat: User CRUD 비밀번호 검증 로직 추가

+++ src/main/java/com/github/gajicoding/schedule_api_project/api/v1/service/impl/UserServiceImpl.java
@@ -1,10 +1,8 @@
 package com.github.gajicoding.schedule_api_project.api.v1.service.impl;
 
 import com.github.gajicoding.schedule_api_project.common.security.PasswordEncryptor;
-import com.github.gajicoding.schedule_api_project.api.v1.data.dto.user.UserLoginRequestDTO;
 import com.github.gajicoding.schedule_api_project.api.v1.data.dto.user.UserRequestDTO;
 import com.github.gajicoding.schedule_api_project.api.v1.data.dto.user.UserResponseDTO;
-import com.github.gajicoding.schedule_api_project.api.v1.data.dto.user.UserSignUpRequestDTO;
 import com.github.gajicoding.schedule_api_project.api.v1.data.entity.User;
 import com.github.gajicoding.schedule_api_project.api.v1.exception.factory.UserExceptionFactory;
 import com.github.gajicoding.schedule_api_project.api.v1.repository.UserRepository;
@@ -14,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,7 +22,19 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserResponseDTO create(UserRequestDTO requestDTO) {
-        return new UserResponseDTO(userRepository.save(requestDTO.toEntity()));
+        // FIXME: AuthService 의 signup 메서드와 중복
+        Optional<User> existingUser = userRepository.findUserByEmail(requestDTO.getEmail());
+
+        // 이메일 중복 체크
+        if(existingUser.isPresent()) {
+            throw UserExceptionFactory.emailAlreadyExists(requestDTO.getEmail());
+        }
+
+        User user = requestDTO.toEntity();
+
+        // 비밀번호 암호화 후 저장
+        user.setPassword(passwordEncryptor.encode(user.getPassword()));
+        return new UserResponseDTO(userRepository.save(user));
     }
 
     @Override
@@ -41,16 +52,30 @@ public class UserServiceImpl implements UserService {
 
     @Transactional
     @Override
-    public UserResponseDTO update(Long id, UserRequestDTO requestDTO) {
+    public UserResponseDTO updateName(Long id, UserRequestDTO requestDTO) {
         User user = userRepository.findById(id).orElseThrow(() -> UserExceptionFactory.notFoundById(id)); // 유저 체크
-        user.update(requestDTO);
 
+        // 비밀번호 일치 확인
+        if(!passwordEncryptor.matches(requestDTO.getPassword(), user.getPassword())) {
+            throw UserExceptionFactory.invalidPassword();
+        }
+
+        // 이름 변경
+        user.setName(requestDTO.getName());
+
+        userRepository.flush(); // 변경 사항 강제 반영
         return new UserResponseDTO(user);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, UserRequestDTO requestDTO) {
         User user = userRepository.findById(id).orElseThrow(() -> UserExceptionFactory.notFoundById(id)); // 유저 체크
+
+        // 비밀번호 일치 확인
+        if(!passwordEncryptor.matches(requestDTO.getPassword(), user.getPassword())) {
+            throw UserExceptionFactory.invalidPassword();
+        }
+
         userRepository.delete(user);
     }
 }
